Added WinRAR to the list of tools you can used and added it to one of the steps.
I don't use UltraISO, therefore, I don't much about the "Extract To..." option, therefore, I've listed WinRAR as a helpful resource as I do have experience with the software. If this addition to the tutorial is problematic or does not makes sense, please open an issue in the issue tracker.